requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart==0.0.6
requests>=2.31.0
pytz>=2023.3
psycopg2-binary==2.9.9
